+++ packages/nimble-tokens/source/styledictionary/config.js
@@ -236,6 +236,19 @@ StyleDictionary.registerTransform({
     }
 });
 
+// Workaround as name/dsp/kebab does not support prefixes
+// See: https://github.com/AdobeXD/design-system-package-dsp/issues/27
+StyleDictionary.registerTransform({
+    name: 'name/nimble/kebab',
+    type: 'name',
+    filter: function (_prop) {
+        return true;
+    },
+    transform: function (prop) {
+        return `ni-nimble-base-${kebabCase(getTokenNameFromProperty(prop))}`;
+    }
+});
+
 StyleDictionary.registerTransform({
     name: 'name/dsp/camel',
     type: 'name',
@@ -269,37 +282,61 @@ StyleDictionary.registerTransform({
     }
 });
 
-// Override default transform groups with custom name transform
-StyleDictionary.registerTransformGroup({
-    name: 'css',
-    transforms: [
-        'attribute/cti',
-        'name/dsp/kebab', // replaces 'name/cti/kebab',
-        'time/seconds',
-        'html/icon',
-        'size/rem',
-        'color/css'
-    ]
+StyleDictionary.registerTransform({
+    type: 'value',
+    transitive: true,
+    name: 'font/weight',
+    filter: token => token.attributes.category === 'font',
+    transform: token => {
+        if (token.value === 'Light') {
+            token.value = '300';
+        } else if (token.value === 'Regular') {
+            token.value = '400';
+        } else if (token.value === 'Semibold') {
+            token.value = '600';
+        } else if (token.value === 'Bold') {
+            token.value = '700';
+        }
+        return token.value;
+    }
 });
 
+// Combination of DSP & Nimble transform overrides
 StyleDictionary.registerTransformGroup({
-    name: 'scss',
+    name: 'nimble/css',
     transforms: [
         'attribute/cti',
-        'name/dsp/kebab', // replaces 'name/cti/kebab',
+        'name/nimble/kebab', // replaces name/dsp/kebab from DSP config
         'time/seconds',
         'html/icon',
-        'size/rem',
+        'size/px', // replaces size/rem from DSP config
         'color/css',
+        'font/weight'
     ]
 });
 
+// Combination of DSP & Nimble transform overrides
 StyleDictionary.registerTransformGroup({
-    name: 'js',
+    name: 'nimble/scss',
+    transforms: [
+        'attribute/cti',
+        'name/nimble/kebab', // replaces name/dsp/kebab from DSP config
+        'time/seconds',
+        'html/icon',
+        'size/px', // replaces size/rem from DSP config
+        'color/css',
+        'font/weight'
+    ]
+});
+
+// Combination of DSP & Nimble transform overrides
+StyleDictionary.registerTransformGroup({
+    name: 'nimble/js',
     transforms: [
         'attribute/cti',
         'name/dsp/pascal', // replaces 'name/cti/pascal',
-        'size/rem',
+        'size/px', // replaces size/rem from DSP config
         'color/hex',
+        'font/weight'
     ]
 });
